2024-12-1 마무리
BoardController / BoardService/ BoardRepository 수정
엔터티 수정

+++ src/main/kotlin/com/galashow/gala/model/entity/Board.kt
@@ -16,7 +16,7 @@ class Board(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "board_no", nullable = false)
-    val boardNo: Int? = null,
+    val boardNo: Long? = null,
 
     @Size(max = 50)
     @NotNull
